pyproject.toml
[tool.poetry]
name = "crewai-code-repo-analyzer"
version = "0.1.0"
description = "Analyze a GitHub code repository and generate documentation for the project."
authors = ["Tim Kitchens <timkitch@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = {extras = ["tools"], version = "^0.30.11"}
pydantic = "^2.6.3"
load-dotenv = "^0.1.0"
crewai-tools = ">=0.2.6,<0.3.0"
plantweb = "^1.2"
langchain_community = "^0.0.29"
langchain_anthropic = "^0.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"